refactor - move code around
